Move mapping to resources and remove duplicate example
